4.17-4.21 && 4.6-4.7

+++ part4/bloglist/controllers/blogs.js
@@ -1,26 +1,56 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { tokenExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
 
+    request.token = tokenExtractor(request)
+    // eslint-disable-next-line no-undef
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
     const blog = new Blog ({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes ? body.likes : 0
+        likes: body.likes ? body.likes : 0,
+        user: user._id
     })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', async (request, response) => { //Blogs isn't deleted in users :(
+    const blog = await Blog.findById(request.params.id)
+    const user = await User.findById(blog.user._id)
+
+    console.log(blog)
+    console.log(user)
+
+    request.token = tokenExtractor(request)
+    // eslint-disable-next-line no-undef
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id || !(blog.user.toString() === user.id.toString()) ) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end
 })
